ADD
Tested frontend making an API request to the backend by adding and using axios on the window. Since cookies are added to the request from the frontend, login was not a problem. This test was done via the browser instead of a third party rest application, i.e. Postman, in order to avoid OAuth issues...the send to async function was created for the user to send a request and response to sendGrid. SendGrid responds by sending email to recipient with custom/modified links. When the user clicks on those links, the server will receive user information that lets the app know which recipient filled out the survey...Dashboard component was tested...materialize button added.